Renamve token to token_detail
